handle /dev/null

+++ src/index.ts
@@ -91,12 +91,18 @@ async function run() {
 
         diff.applyPatches(patchContent, {
           loadFile(index: diff.ParsedDiff, callback: (err: any, data: string) => void) {
-            
+
             let oldFileName = index.oldFileName!;
             // replace 'a' with actual repository path
             oldFileName = oldFileName.replace(/a\//, path);
 
             console.log(`Loading file ${oldFileName}`);
+
+            if (oldFileName === "/dev/null") {
+              callback(null, "");
+              return;
+            }
+
             callback(null, fs.readFileSync(oldFileName, "utf8"));
           },
           patched(index: diff.ParsedDiff, content: string, callback: (err: any, data: string) => void) {
@@ -108,12 +114,25 @@ async function run() {
             // replace 'a' with actual repository path
             oldFileName = oldFileName.replace(/a\//, path);
 
-            console.log(`Patching file ${index.oldFileName}`);
-            fs.writeFileSync(newFileName, content);
-            if (oldFileName !== newFileName) {
+            console.log(`Patching file: new file name - ${newFileName}, old file name - ${oldFileName}`);
+
+            if (newFileName === "/dev/null") {
+              console.log(`Deleting file ${oldFileName}`);
+              fs.unlinkSync(oldFileName);
+              callback(null, content);
+              return;
+            }
+
+            if (oldFileName !== newFileName && oldFileName !== "/dev/null") {
               console.log(`Renaming file ${oldFileName} to ${newFileName}`);
               fs.unlinkSync(oldFileName!);
+              fs.writeFileSync(newFileName, content);
+              callback(null, content);
+              return;
             }
+
+            console.log(`Writing to file ${newFileName}`);
+            fs.writeFileSync(newFileName, content);
             callback(null, content);
           },
           complete(err: any) {
